+++ portal/www/amstatus.php
@@ -34,18 +34,8 @@ show_html_head('GENI Aggregate Status');
 <div id="content-outer" class="one-card">
 <div id="content">
 <h1>GENI Aggregate Status</h1>
-  <section>
-    The following is a summary up/down status for each aggregate as reported by
-    <a href="https://genimon.uky.edu/login" target="_blank">
-        GENI Monitoring</a>.
-    For detailed information go to the
-    <a href="https://genimon.uky.edu/status" target="_blank">
-        GENI monitoring aggregate status page</a> (no login required).
-  </section>
-  <section>
-    <h3>Legend</h3>
-    <?php print agg_status_legend(); ?>
-  </section>
+  <?php print agg_status_monitoring_section(); ?>
+  <?php print agg_status_legend_section(); ?>
   <section>
     <h3>Aggregate Status</h3>
     <div class='tablecontainer'>